@@ -11,18 +11,18 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.2",
+        "php": "^8.2",
         "neomerx/json-api": "^0.8",
-        "doctrine/common": "^2.6 || ^3.0",
-        "symfony/property-access": "^4.0 || ^5.0",
-        "symfony/validator": "^4.0 || ^5.0",
-        "symfony/http-kernel": "^4.0 || ^5.0"
+        "doctrine/common": "^3.0",
+        "symfony/property-access": "^7.0",
+        "symfony/validator": "^7.0",
+        "symfony/http-kernel": "7.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.2",
         "scrutinizer/ocular": "^1.3",
-        "squizlabs/php_codesniffer": "^2.5",
-        "doctrine/orm": "^2.5"
+        "squizlabs/php_codesniffer": "^3.10",
+        "doctrine/orm": "^3.2"
     },
     "autoload": {
         "psr-4": {
