@@ -2,29 +2,30 @@
 
 namespace Reva2\JsonApi\Tests\Fixtures\Metadata;
 
-use Reva2\JsonApi\Annotations\ApiObject;
-use Reva2\JsonApi\Annotations\Property;
+use Reva2\JsonApi\Attributes as API;
 
 /**
  * @ApiObject()
  */
+#[API\ApiObject()]
 class OfficesListMetadata
 {
     /**
      * @var string
-     * @Property(type="integer", path="inner_int.one")
      */
+    #[API\Property(type: "integer", path: "inner_int.one")]
     public $someInnerIntOne;
 
     /**
      * @var int
      * @Property(type="integer", path="inner_int.two")
      */
+    #[API\Property(path: "inner_int.two")]
     public $someInnerIntTwo;
 
     /**
      * @var int
-     * @Property(type="integer")
      */
-    public $someInt;
+    #[API\Property]
+    public int $someInt;
 }
