@@ -8,36 +8,32 @@
  * file that was distributed with this source code.
  */
 
-
 namespace Reva2\JsonApi\Tests\Fixtures\Resources;
 
-use Reva2\JsonApi\Annotations\ApiResource;
-use Reva2\JsonApi\Annotations\Attribute;
-use Reva2\JsonApi\Annotations\Id;
+use Reva2\JsonApi\Attributes as API;
 
 /**
  * Something resource
  *
  * @author Sergey Revenko <sergey.revenko@orbitsoft.com>
  * @package Reva2\JsonApi\Tests\Fixtures\Resources
- *
- * @ApiResource(name="something")
  */
+#[API\Resource('something')]
 class Something
 {
     /**
      * Resource id
      *
      * @var string
-     * @Id()
      */
-    public $id;
+    #[API\Id]
+    public string $id;
 
     /**
      * Resource name
      *
      * @var string
-     * @Attribute()
      */
-    public $name;
+    #[API\Attribute]
+    public string $name;
 }
