@@ -11,7 +11,7 @@
 
 namespace Reva2\JsonApi\Tests\Fixtures\Objects;
 
-use Reva2\JsonApi\Annotations as API;
+use Reva2\JsonApi\Attributes as API;
 
 /**
  * Another invalid JSON API object
@@ -23,7 +23,7 @@ class InvalidObject2
 {
     /**
      * @var mixed
-     * @API\Property(parser="parseData")
      */
+    #[API\Property(parser: "parseData")]
     public $data;
 }
