@@ -11,22 +11,21 @@
 
 namespace Reva2\JsonApi\Tests\Fixtures\Resources;
 
-use Reva2\JsonApi\Annotations as API;
+use Reva2\JsonApi\Attributes as API;
 
 /**
  * Example JSON API resource that represent dog
  *
  * @package Reva2\JsonApi\Tests\Fixtures\Resources
  * @author Sergey Revenko <dedsemen@gmail.com>
- *
- * @API\ApiResource()
  */
+#[API\Resource]
 class Dog extends Pet
 {
     /**
      * @inheritdoc
      */
-    public function whoIAm()
+    public function whoIAm(): string
     {
         return 'dog';
     }
